fix addKid + mv sanitize
