+++ src/routes/sanitize/+server.ts
@@ -1,11 +1,5 @@
-import { PrismaClient, Pronoun, type User } from '@prisma/client';
 import { json } from '@sveltejs/kit';
-const prisma = new PrismaClient();
-import sanitizerFunc from 'sanitize';
-
-const sanitizer = sanitizerFunc();
-
-type PRONOUNS_ENUM = keyof typeof Pronoun;
+import { getParams, circleNotif, dateNotes } from '$lib/server/sanitize';
 
 const WHICH_PARAMS: { [key: string]: string[] } = {
 	circleNotif: ['user', 'schedDiffs'],
@@ -29,47 +23,3 @@ export async function GET({ url }: { url: URL }) {
 
 	return new Response(JSON.stringify({ message: "Unknown 'which'" }), { status: 400 });
 }
-
-const getParams = (url: URL, paramNames: string[]) => {
-	const params = paramNames.map((x) => url.searchParams.get(x)).filter(Boolean) as string[];
-	if (params.length < paramNames.length) return null;
-	return params;
-};
-
-const sanitize = (input: string) => sanitizer.value(input, 'str');
-
-const circleNotif = async (userStr: string, schedDiffs: string) => {
-	const user = JSON.parse(userStr) as User;
-	const sanitizedSchedDiffs = sanitize(schedDiffs);
-	let objectivePronoun = Pronoun[user.pronouns as PRONOUNS_ENUM].split('_')[2];
-	const { SHE_HER_HERS, THEY_THEM_THEIRS, XE_XEM_XYRS, ZEZIE_HIR_HIRS } = Pronoun;
-	// turn from possessive noun to possessive adjective
-	switch (user.pronouns) {
-		case SHE_HER_HERS:
-		case THEY_THEM_THEIRS:
-		case XE_XEM_XYRS:
-		case ZEZIE_HIR_HIRS:
-			objectivePronoun = objectivePronoun.slice(0, -1).toLowerCase();
-	}
-
-	let kidNames: string = '';
-
-	if (!user.householdId) {
-		return new Response(JSON.stringify({ message: 'null householdId' }), { status: 400 });
-	}
-	const kids = await prisma.householdChild.findMany({
-		where: {
-			householdId: user.householdId
-		}
-	});
-
-	kidNames = kids
-		.map((kid) => `${kid.firstName}${kid.lastName ? ` ${kid.lastName}` : ''}`)
-		.join(', ');
-
-	return `${user.firstName}${
-		user.lastName && user.lastName.length ? ` ${user.lastName}` : ''
-	} (parent of ${kidNames}) has updated ${objectivePronoun} tentative schedule:\nLegend: 🏠(host) 🚗(visit) 👤(dropoff) 👥(together) 🏫(at school) ⭐(good) 🌟(great) 🙏(needed)\n\n${sanitizedSchedDiffs}`;
-};
-
-const dateNotes = (notes: string) => sanitize(notes);
